NarrowSearchAI/requirements.txt
streamlit
pandas
pyarrow
plotly
pydantic
sentence-transformers
faiss-cpu
scikit-learn
loguru
rich
